requirements.txt
SpeechRecognition==3.15.1
PyAudio==0.2.14
pyttsx3==2.90
google-generativeai==0.8.5
groq==0.18.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
PyQt6==6.8.1
send2trash==1.8.3
psutil==7.0.0
pyperclip==1.9.0
pyautogui==0.9.54
pathlib==1.0.1
requests==2.32.3
sounddevice>=0.4.6
scipy>=1.11.0
